Avoid deleting tab locations when tmux isn't running

+++ Tools/my-tools/tmux-kill-session.sh
@@ -16,5 +16,7 @@
 #
 ################################################################################
 
-tmux list-windows -F '#{pane_current_path}' > ~/Tools/my-tools/.tmux-saved-tabs.txt
-tmux kill-session
+if [ "$TERM" = "screen" ] && [ -n "$TMUX" ]; then
+	tmux list-windows -F '#{pane_current_path}' > ~/Tools/my-tools/.tmux-saved-tabs.txt
+	tmux kill-session
+fi
